+++ pkg/crypto/hashing/hashing.go
@@ -2,6 +2,7 @@ package hashing
 
 import (
 	"crypto/sha512"
+	"sync"
 
 	"github.com/number571/go-peer/pkg/encoding"
 )
@@ -15,19 +16,20 @@ const (
 )
 
 type sSHA512Hasher struct {
+	fOnce    sync.Once
 	fHash    []byte
 	fHashStr string
 }
 
 func NewHasher(pData []byte) IHasher {
 	s := sha512.Sum384(pData)
-	return &sSHA512Hasher{
-		fHash:    s[:],
-		fHashStr: encoding.HexEncode(s[:]),
-	}
+	return &sSHA512Hasher{fHash: s[:]}
 }
 
 func (p *sSHA512Hasher) ToString() string {
+	p.fOnce.Do(func() {
+		p.fHashStr = encoding.HexEncode(p.ToBytes())
+	})
 	return p.fHashStr
 }