+++ pkg/crypto/hashing/hmac.go
@@ -3,6 +3,7 @@ package hashing
 import (
 	"crypto/hmac"
 	"crypto/sha512"
+	"sync"
 
 	"github.com/number571/go-peer/pkg/encoding"
 )
@@ -12,6 +13,7 @@ var (
 )
 
 type sHMACSHA512Hasher struct {
+	fOnce    sync.Once
 	fHash    []byte
 	fHashStr string
 }
@@ -20,13 +22,13 @@ func NewHMACHasher(pKey []byte, pData []byte) IHasher {
 	h := hmac.New(sha512.New384, pKey)
 	h.Write(pData)
 	s := h.Sum(nil)
-	return &sHMACSHA512Hasher{
-		fHash:    s,
-		fHashStr: encoding.HexEncode(s),
-	}
+	return &sHMACSHA512Hasher{fHash: s}
 }
 
 func (p *sHMACSHA512Hasher) ToString() string {
+	p.fOnce.Do(func() {
+		p.fHashStr = encoding.HexEncode(p.ToBytes())
+	})
 	return p.fHashStr
 }